requirements.txt
Django>=3.2.9,<4.0
Djangorestframework>=3.12.4,<4.0
psycopg2-binary>=2.9.2
